Update the molecule converge configurations
Update the converge configurations for our molecule scenarios to match the
format that molecule uses now. This matches the format of a `converge.yml`
file if one were to run `molecule init role <role name> --driver-name docker`
with a current version of molecule. This change is necessary to allow these
configurations to pass `ansible-lint` now that we're upgrading to v5.