Add Debian 11 (Bullseye) support
This adds support for Debian Bullseye to this Ansible role. This includes
changing the version of PowerShell installed from 7.1.2 to 7.1.5 (the latest)
This is necessary because 7.1.2 is not installable on Debian Bullseye due to
an unfulfillable requirement. This was resolve in version 7.1.3 and beyond.
This can be seen by comparing the dependencies of the packages.

v7.1.2:
Depends: libc6, libgcc1, libgssapi-krb5-2, liblttng-ust0, libstdc++6, zlib1g,
         libssl1.1, libicu63
v7.1.3:
Depends: libc6, libgcc1, libgssapi-krb5-2, liblttng-ust0, libstdc++6, zlib1g,
         libicu72|libicu71|libicu70|libicu69|libicu68|libicu67|libicu66|libicu65|libicu63|libicu60|libicu57|libicu55|libicu52,
         libssl1.1|libssl1.0.2|libssl1.0.0
v7.1.5:
Depends: libc6, libgcc1, libgssapi-krb5-2, liblttng-ust0, libstdc++6, zlib1g,
         libicu72|libicu71|libicu70|libicu69|libicu68|libicu67|libicu66|libicu65|libicu63|libicu60|libicu57|libicu55|libicu52,
         libssl1.1|libssl1.0.2|libssl1.0.0

The root cause was that the libicu63 package is not available for Debian
Bullseye, but 7.1.3 and later expanded the range of supported libicu versions.
This includes the version available on Debian Bullseye (libicu67).