Change the project installed in the converge playbook of the rust scenario
CrackMapExec is the reason this functionality is being added, since it
is a Python tool with no wheel available that requires a Rust compiler
to build its native code.  Unfortunately, CrackMapExec does not make a
good test case.  The reason is that some of its dependencies are not
versioned but instead just pull in the HEAD of the default branch.
Such an installation can never pass idempotence.

Since the new functionality allows for the building of a Rust-based
project is makes sense to go ahead and test that use case.