Add a role variable that can force the installation of Python
This role variable also forces the creation of an empty Python virtual
environment.  This is occassionally useful, e.g., when installing a
tool that lacks any Python packaging and has no dependencies.

We also add a test case to the "Python" Molecule scenario that tests
the use of this new variable.